week8/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
aiostream==0.5.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dataclasses-json==0.6.7
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
docx2txt==0.8
eval_type_backport==0.2.2
exceptiongroup==1.3.0
fake-useragent==1.4.0
frozenlist==1.7.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
googlesearch-python==1.2.3
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib-metadata==6.11.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpath-python==1.0.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
llama-index==0.10.43
llama-index-agent-openai==0.2.9
llama-index-cli==0.1.13
llama-index-core==0.10.43
llama-index-embeddings-mistralai==0.1.4
llama-index-embeddings-openai==0.1.11
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-legacy==0.9.48.post4
llama-index-llms-mistralai==0.1.6
llama-index-llms-openai==0.1.26
llama-index-multi-modal-llms-openai==0.1.9
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.33
llama-index-readers-llama-parse==0.1.6
llama-parse==0.4.9
llamaindex-py-client==0.1.19
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mistralai==0.4.2
multidict==6.6.3
mypy_extensions==1.1.0
narwhals==1.48.1
nest-asyncio==1.5.8
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
openai==1.97.1
orjson==3.10.18
outcome==1.3.0.post0
packaging==23.2
pandas==2.3.1
pillow==10.4.0
propcache==0.3.2
protobuf==4.25.8
pyarrow==15.0.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pypdf==4.3.1
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.31.0
rich==13.9.4
rpds-py==0.26.0
selenium==4.15.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.41
streamlit==1.28.1
striprtf==0.0.26
tenacity==8.5.0
tiktoken==0.9.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
validators==0.35.0
watchdog==6.0.0
webdriver-manager==4.0.1
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.1
zipp==3.23.0
